Move the key related config parser into a separate file
